Support all three kinds of formals in PlainLambda
JavaScript has `arguments` keywords to do varargs thing.  We generate
bunch of VarDecs for cases where we do `(id...+ . v)` and `id` cases.
The `arguments` object in JS is not exactly Array, so we convert that to
proper array first followed by converting that Racket list.

In general cases, arguments object is not used at all.