ffi: unwrap Symbol in ILRef/ILIndex
Static symbol are wrapped inside ILValue, making
it compile down to Rapture Symbol object. We don't
want that in this case.

In case of ILRef/ILVal, a Symbol lvalue actually
means an identifier.

The reason it gets wrapped around ILValue and not
treated as id, is because to Racket all FFI feature
appear as function (a hack), so it gets quoted.