expand: Import modules as objects
Instead of importing individual identifiers from JS modules, we import
the whole module as an object and ref through it to access individual
exports. There are two kinds of modules involved, one being original
module where it is defined, and other being from where it is being
imported. We ref ident from original module, but still like to import
the other module from which it is re-exported to execute any possible
side-effects there.

This also solves the problem where we get same identifier name, with
different bindings because of macro expansion.  Since these idents might
not be actually provided, we keep track of these and  export them in
final JS.

To keep track of where an identifier come from, from now on ident is no
longer a simple symbol in absyn, but a union of different kind, such
ImportedIdent, LexicalIdent...