compiler: For generating stubs only export what is implemented
Earlier we looked into the the `used-idents` dictionary to see what
identifier are used for that module and export those. This was
problamatic because it will add code for the identifiers we have not
actually implemented and JavaScript will run until it realized that the
identifier is undefined. Now we look into JS file directly for exports.