js-support: fix hamt library copy and import
NodeJs style modules don't work well with traceur, so we explicitely
appended an ES6 export. However, since the previous export statements
were still there, it failed in case of babel target because of the
mixup. We remove all legacy export code via gulpfile, and replace with
ES6 export.

This bug was undiscovered because we didn't hash library in babel
environment, until we added parameterization which uses it at
initialization. (ISSUE #6)