Re-implement hashing for better performance
This implementation is cleaner and considerably faster than the previous
one, as per the microbenchmarks linked in the comments. For example,
`hashNumber` is **175x faster**.

The previous code was slow as a result of supporting browsers that are
now obsolete, and had branches that were no longer hit after the hashing
overhaul in the Unicode PR.

The `smi` function has been replaced with `~~h` because on 64-bit platforms
it's not necessary, as the SMI size is 32 bits:
https://github.com/v8/v8/blob/fdfb8c9efbbd96cde5e4cc09cd22146d75e5598b/include/v8.h#L8773

Since all the Facebook (ImmutableJS) code has been removed, this repo is
no longer encumbered by the Facebook's PATENTS clause, which made using
it impossible or difficult in a lot of companies.