Updates how Symbols are printed/displays in a variety of locations.
JS Symbols print like "Symbol(foo)" when .toString is called on
them. To correctly pass the integration tests and mimic Racket's
printing behavior we need to pull the value from interned Symbols and
display that instead.

This commit adds some utils to make it easier to check if a value is a
Symbol and correctly return the type of string we want to use in display.